replace ProcessingStep stack with batch job stack
